Move the client in its dedicated package
